cleaned up splits view and average pace calculation.
