+++ Olympsis/Data/Global Variables.swift
@@ -6,8 +6,8 @@
 //
 
 import UIKit
-import Foundation
 import SwiftUI
+import Foundation
 
 let SCREEN_WIDTH = UIScreen.main.bounds.width
 let SCREEN_HEIGHT = UIScreen.main.bounds.height
