--- Olympsis/Shared/Components/WorkoutMapView.swift
@@ -1,34 +1,0 @@
-//
-//  RunningMapView.swift
-//  Olympsis
-//
-//  Created by Joel Joseph on 7/6/24.
-//
-
-import os
-import MapKit
-import SwiftUI
-import HealthKit
-
-struct WorkoutMapView: View {
-    
-    @State var locations: [CLLocationCoordinate2D] = RUNNING_POINTS
-    
-    var body: some View {
-        Map {
-            MapPolyline(coordinates: locations)
-                .tint(Color.blue)
-                .stroke(Color.blue, lineWidth: 5)
-            
-        }
-        .clipShape(RoundedRectangle(cornerRadius: 10))
-        .padding(.horizontal)
-        .frame(height: 300)
-    }
-}
-
-#Preview {
-    WorkoutMapView()
-}
-
-
