--- Olympsis/Shared/Data/WorkoutManager+Write.swift
@@ -1,15 +1,0 @@
-//
-//  WorkoutManager+Write.swift
-//  Olympsis
-//
-//  Created by Joel Joseph on 8/3/24.
-//
-
-import os
-import SwiftUI
-
-extension WorkoutManager {
-    func startRecordingLocation() {
-        
-    }
-}
