+++ Shared/Data/WorkoutManager+Splits.swift
@@ -72,7 +72,6 @@ extension WorkoutManager {
                     segmentNumber: splits.count + 1,
                     distance: splitDistance,
                     duration: duration,
-                    pace: calculatedPace,
                     elevationGain: elevationData.gain,
                     elevationLoss: elevationData.loss,
                     startTime: splitStartTime,
@@ -150,7 +149,6 @@ extension WorkoutManager {
                     segmentNumber: splits.count + 1,
                     distance: splitDistance,
                     duration: duration,
-                    pace: calculatedPace,
                     elevationGain: elevationData.gain,
                     elevationLoss: elevationData.loss,
                     startTime: splitStartTime,
@@ -190,7 +188,6 @@ extension WorkoutManager {
                 segmentNumber: splits.count + 1,
                 distance: currentDistance,
                 duration: duration,
-                pace: duration / (currentDistance / (unit == UnitLength.miles ? 1609.344 : 1000.0)),
                 elevationGain: elevationData.gain,
                 elevationLoss: elevationData.loss,
                 startTime: splitStartTime,
@@ -264,7 +261,6 @@ extension WorkoutManager {
                 segmentNumber: splitIndex + 1,
                 distance: splitDistanceInMeters,
                 duration: splitDuration,
-                pace: averagePacePerMeter * (unit == UnitLength.miles ? 1609.344 : 1000.0),
                 elevationGain: 0,
                 elevationLoss: 0,
                 startTime: startTime,
